% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_3_function_3.R
\name{my_bayesian}
\alias{my_bayesian}
\title{Project 3: Function 3 my_bayesian}
\usage{
my_bayesian(df)
}
\arguments{
\item{df}{data frame}
}
\value{
model summary for both classical and Bayesian regression, including trace and history plots and the convergence diagnostic
}
\description{
Project 3: Function 3 my_bayesian
}
\examples{
\dontrun{my_bayesian(df = prodqual)}
}
