% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_3_function_2.R
\name{my_bootstrap}
\alias{my_bootstrap}
\title{Project 3: Function 2 my_bootstrap}
\usage{
my_bootstrap(df, iter, alpha)
}
\arguments{
\item{df}{dataframe}

\item{iter}{iteration time}

\item{alpha}{significance level}
}
\value{
point estimate from both mlr and bootstrapping
}
\description{
Project 3: Function 2 my_bootstrap
}
\examples{
\dontrun{my_bootstrap(df = prodqual, iter = 10000, alpha = 0.05)}
}
