% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers_test.R
\name{my_fishser}
\alias{my_fishser}
\title{Fisher's test}
\usage{
my_fishser(filename)
}
\arguments{
\item{filename}{filename name of the data file}
}
\value{
table chi_squared_test fishers_test
}
\description{
Fisher's test
}
\examples{
\dontrun{my_fishser("BONDING")}

}
