% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_2_model_plot.R
\name{my_model_plot}
\alias{my_model_plot}
\title{Model plotting}
\usage{
my_model_plot(dataframe)
}
\arguments{
\item{dataframe}{name of the data frame}
}
\value{
model summaries and plots of both full sample and sub samples
}
\description{
Model plotting
}
\examples{
\dontrun{my_model_plot(fuels)}

}
