% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way.R
\name{my_one_way}
\alias{my_one_way}
\title{One Way Table Analysis}
\usage{
my_one_way(filename, alpha = 0.05)
}
\arguments{
\item{filename}{name of the data file}

\item{alpha}{significance level}
}
\value{
table, frequency
}
\description{
One Way Table Analysis
}
\examples{
\dontrun{my_one_way("SCAN")}
}
