% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_2_residual.R
\name{my_residual}
\alias{my_residual}
\title{Residual Analysis}
\usage{
my_residual(dataframe)
}
\arguments{
\item{dataframe}{name of the data frame}
}
\value{
analysis of MLR assumptions regarding to the error term
}
\description{
Residual Analysis
}
\examples{
\dontrun{my_residual(fuels)}

}
