% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way.R
\name{my_two_way}
\alias{my_two_way}
\title{Two Way Table Analysis}
\usage{
my_two_way(filename, alpha = 0.05)
}
\arguments{
\item{filename}{name of the data file}

\item{alpha}{significance level}
}
\value{
table, chisq
}
\description{
Two Way Table Analysis
}
\examples{
\dontrun{my_one_way("HYBRID")}
}
