% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_3_function_1.R
\name{my_validity}
\alias{my_validity}
\title{Project 3: Function 1 my_validity}
\usage{
my_validity(dataframe)
}
\arguments{
\item{dataframe}{df}
}
\value{
model summary, reponse surface, residual-fitted value plot, scale-location plot, constant variance test result, shapiro test result
}
\description{
Project 3: Function 1 my_validity
}
\examples{
\dontrun{my_validity(dataframe = prodqual)}
}
