% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{myboot}
\alias{myboot}
\title{Bootstrapping estimates}
\usage{
myboot(df, model, iter, alpha)
}
\arguments{
\item{df}{name of the data frame}

\item{model}{model specification}

\item{iter}{iteration times}

\item{alpha}{significance level}
}
\value{
invisible named list containing a data frame of iterations, the CIs, and point estimates
}
\description{
Bootstrapping estimates
}
\examples{
\dontrun{myboot(df = bubble, model = "model1", iter = 1000, alpha = .95)}

}
