% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_fit.R
\name{mypiece_fit}
\alias{mypiece_fit}
\title{Piecewise_fit}
\usage{
mypiece_fit(data)
}
\arguments{
\item{data}{name of the data set}
}
\value{
plots of r squared and aic
}
\description{
Piecewise_fit
}
\examples{
\dontrun{mypiece_fit(data = ddt)}
}
