% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_optimal.R
\name{mypiece_optimal}
\alias{mypiece_optimal}
\title{Piecewise_optimal}
\usage{
mypiece_optimal(data)
}
\arguments{
\item{data}{name of the data set}
}
\value{
x_k y_k r_sq aic
}
\description{
Piecewise_optimal
}
\examples{
 \dontrun{mypiece_optimal(data = ddt)}

}
