% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of Tickets for Overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability of overbooking}

\item{p}{probability of each passenger showing up}
}
\value{
plots of objective function using discrete and continuous distributions
}
\description{
Number of Tickets for Overbooking
}
\examples{
\dontrun{ntickets(N=200, gamma = 0.02, p = 0.95)}
}
