% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_3_function_4.R
\name{shiny_project3}
\alias{shiny_project3}
\title{Title Dynamic plots using shiny}
\usage{
shiny_project3()
}
\value{
Plots and summary
}
\description{
Title Dynamic plots using shiny
}
\examples{
\dontrun{shiny_project3()}
}
