% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Binomial Application to Overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on a flight}

\item{gamma}{Probability that more people show up than there are seats on the plane}

\item{p}{Probability a ticketholder shows up for the flight}
}
\value{
The optimal number of tickets sold based on the discrete binomial distribution, denoted by nd, and the continuous normal approximation, denoted by nc.
}
\description{
Calculates the number of tickets to be sold for a flight using the binomial distribution and the normal approximation of the binomial distribution. Creates plots of the number of tickets sold against the objective function to find the optimal ticket sales. Prints optimal ticket values.
}
\examples{
ntickets(400, 0.02, 0.95)


}
